tooltip: Fix possibly briefly appearing at 0x0
When showing a tooltip on the edge of a monitor, the tooltip could be wrongly
placed and be shown going from one monitor to the next.

This happened because the allocation for current_window might have been reset
when hidden, and both width & height were set to 1, hence wrong calculations.

Making the window visible could lead to tooltip briefly appearing at 0x0, so
instead the natural size is requested & used.

https://bugzilla.gnome.org/show_bug.cgi?id=698730
Signed-off-by: Olivier Brunel <jjk@jjacky.com>
Signed-off-by: Michael Webster <miketwebster@gmail.com>