Banshee:  build: linking against dbus-glib-1 is now mandatory (bgo#692374)
The work-around for this bug was not applied successfully by some
packagers because we didn't make the dependency mandatory. Now that
gsettings branch is almost ready to be merged to master (which
means this workaround would be reverted), it's good if we make sure
that at least the last GConf-enabled versions don't suffer this bug.

https://github.com/GNOME/banshee/commit/5e2c69b659c1b5642d149ce8a61ca2d6c3fd5605