Banshee: ThickClient: prevent Select All/None menu options disappearing (bgo#726929)
These options were disappearing from the "Edit" menu when the user
clicked into a content-filtered area (i.e.: album grid) and not
appearing back properly when going back focusing the tracks listview.

In the case FilterFocused is false, we should call UpdateActions() too
but with true params for visible & sensitive.

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>

https://github.com/GNOME/banshee/commit/7527e73383b6a4a258775c7eda03d94fbe3c5f30