Banshee:  ThickClient: fix previous commit wrt Select All/None options (bgo#726929)
Previous commit was wrong because it would override the visible
and sensible properties for non-Select options which were already
properly set at the beginning of the method (i.e. "SearchMenuAction").

We should use UpdateActions(true,true,...) then at the beginning,
so each of the options can be adjusted later if needed.

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>

https://github.com/GNOME/banshee/commit/097be47a3d141ef3f77d3df03854088856c85f47